update 8.1 - by chris
18.1 Fixes & Improvements

05/3/21: Fixed issue on fullfillment sheet where KIT was displayed by mistake
05/1/21: Only update variation selling price if different than regular price when receiving
4/28/21: Change label of table report for reports --> recv. --> cheapest supplier
4/27/21: Added item id to suspended sales report
4/27/21: Added store config option Show Person ID on receipt
4/23/21: Shopify e-commerce now supports main image in php pos
4/23/21: Fixed code deprecation issue on receipts
4/22/21: Use correct url for editing sale from work order
4/22/21: Don't show cost price in grid without permission
4/21/21: Closeout report for store accounts now does NOT include deleted customers/suppliers
4/21/21: New work order's sales are now created as estiamtes so they don't show until payment collected...

+++ application/models/Item_attribute.php
@@ -176,6 +176,22 @@ class Item_attribute extends MY_Model
 		return ($query->num_rows()==1);
 	}
 	
+	function get_attribute_id($attr_name,$item_id = NULL)
+	{
+		$this->db->select('id');
+		$this->db->from('attributes');
+		$this->db->where('name',(string)$attr_name);
+		$this->db->where('item_id',$item_id);
+		$query = $this->db->get();
+
+		if ($query->num_rows()==1)
+		{
+			return $query->row()->id;
+		}
+		
+		return NULL;
+	}
+	
 	function get_attributes_for_ecommerce($attribute_ids = array())
 	{
 		$this->db->from('attributes');
@@ -314,15 +330,28 @@ class Item_attribute extends MY_Model
 		return false;		
 	}
 	
-	function save_item_attributes($attribute_ids, $item_id)
+	function save_item_attributes($attribute_ids, $item_id,$delete = TRUE)
 	{
-		$this->db->delete('item_attributes', array('item_id' => $item_id));
+		if ($delete)
+		{
+			$this->db->delete('item_attributes', array('item_id' => $item_id));
+		}
 		
 		foreach($attribute_ids as $attribute_id)
 		{
-			if(!$this->db->insert('item_attributes',array('attribute_id' => $attribute_id,'item_id' => $item_id)))
+			if ($delete)
 			{
-				return false;
+				if(!$this->db->insert('item_attributes',array('attribute_id' => $attribute_id,'item_id' => $item_id)))
+				{
+					return false;
+				}
+			}
+			else
+			{
+				if(!$this->db->replace('item_attributes',array('attribute_id' => $attribute_id,'item_id' => $item_id)))
+				{
+					return false;
+				}
 			}
 		}
